+++ cmd/frontend/graphqlbackend/search_results.go
@@ -1482,9 +1482,12 @@ func (r *searchResolver) doResults(ctx context.Context, args *search.TextParamet
 
 	args.RepoOptions = r.toRepoOptions(args.Query, resolveRepositoriesOpts{})
 
+	// globalSearch controls whether we run a global zoekt search.
+	globalSearch := args.Mode == search.ZoektGlobalSearch
+
 	// performance optimization: call zoekt early, resolve repos concurrently, filter
 	// search results with resolved repos.
-	if args.Mode == search.ZoektGlobalSearch {
+	if globalSearch {
 		argsIndexed := *args
 
 		userID := int32(0)
@@ -1522,7 +1525,8 @@ func (r *searchResolver) doResults(ctx context.Context, args *search.TextParamet
 				ctx, stream, cleanup := streaming.WithLimit(ctx, agg, int(argsIndexed.PatternInfo.FileMatchLimit))
 				defer cleanup()
 
-				zoektArgs, err := zoekt.NewIndexedSearchRequest(ctx, &argsIndexed, search.TextRequest, zoekt.MissingRepoRevStatus(stream))
+				// This code path implies global-search (3rd arg is true).
+				zoektArgs, err := zoekt.NewIndexedSearchRequest(ctx, &argsIndexed, true, search.TextRequest, zoekt.MissingRepoRevStatus(stream))
 				if err != nil {
 					agg.Error(err)
 					return
@@ -1622,7 +1626,8 @@ func (r *searchResolver) doResults(ctx context.Context, args *search.TextParamet
 				ctx, stream, cleanup := streaming.WithLimit(ctx, agg, int(args.PatternInfo.FileMatchLimit))
 				defer cleanup()
 
-				zoektArgs, err := zoekt.NewIndexedSearchRequest(ctx, args, search.TextRequest, zoekt.MissingRepoRevStatus(stream))
+				// This code path implies we've already decided to run global-search (3rd arg is always false).
+				zoektArgs, err := zoekt.NewIndexedSearchRequest(ctx, args, false, search.TextRequest, zoekt.MissingRepoRevStatus(stream))
 				if err != nil {
 					agg.Error(err)
 					return