+++ enterprise/cmd/frontend/internal/batches/resolvers/batch_spec_test.go
@@ -34,7 +34,7 @@ func TestBatchSpecResolver(t *testing.T) {
 	}
 
 	ctx := actor.WithInternalActor(context.Background())
-	db := dbtest.NewDB(t, "")
+	db := dbtest.NewDB(t)
 
 	cstore := store.New(db, &observation.TestContext, nil)
 	repoStore := database.ReposWith(cstore)
@@ -260,7 +260,7 @@ func TestBatchSpecResolver_BatchSpecCreatedFromRaw(t *testing.T) {
 	}
 
 	ctx := context.Background()
-	db := dbtest.NewDB(t, "")
+	db := dbtest.NewDB(t)
 
 	now := timeutil.Now().Truncate(time.Second)
 	minAgo := func(min int) time.Time { return now.Add(time.Duration(-min) * time.Minute) }
@@ -304,10 +304,8 @@ func TestBatchSpecResolver_BatchSpecCreatedFromRaw(t *testing.T) {
 
 	applyUrl := fmt.Sprintf("/users/%s/batch-changes/apply/%s", admin.Username, apiID)
 	codeHosts := apitest.BatchChangesCodeHostsConnection{
-		TotalCount: 1,
-		Nodes: []apitest.BatchChangesCodeHost{
-			{ExternalServiceKind: "GITHUB", ExternalServiceURL: "https://github.com/"},
-		},
+		TotalCount: 0,
+		Nodes:      []apitest.BatchChangesCodeHost{},
 	}
 	want := apitest.BatchSpec{
 		Typename: "BatchSpec",