+++ enterprise/cmd/frontend/internal/batches/resolvers/bulk_operation_connection_test.go
@@ -25,7 +25,7 @@ func TestBulkOperationConnectionResolver(t *testing.T) {
 	}
 
 	ctx := actor.WithInternalActor(context.Background())
-	db := dbtest.NewDB(t, "")
+	db := dbtest.NewDB(t)
 
 	userID := ct.CreateTestUser(t, db, true).ID
 	now := timeutil.Now()