+++ enterprise/internal/batches/reconciler/reconciler_test.go
@@ -23,7 +23,7 @@ func TestReconcilerProcess_IntegrationTest(t *testing.T) {
 	}
 
 	ctx := actor.WithInternalActor(context.Background())
-	db := dbtest.NewDB(t, "")
+	db := dbtest.NewDB(t)
 
 	store := store.New(db, &observation.TestContext, nil)