+++ internal/actor/actor.go
@@ -65,7 +65,7 @@ func (a *Actor) User(ctx context.Context, fetcher userFetcher) (*types.User, err
 	a.userOnce.Do(func() {
 		a.user, a.userErr = fetcher.GetByID(ctx, a.UID)
 	})
-	if a.user.ID != a.UID {
+	if a.user != nil && a.user.ID != a.UID {
 		return nil, errors.Errorf("actor UID (%d) and the ID of the cached User (%d) do not match", a.UID, a.user.ID)
 	}
 	return a.user, a.userErr